chore: refactor server.js to export app without running server
