+++ app.js
@@ -1,0 +1,35 @@
+import express from 'express';
+import colors from 'colors';
+import dotenv from 'dotenv';
+import morgan from 'morgan';
+import authRoutes from './routes/authRoute.js';
+import categoryRoutes from './routes/categoryRoutes.js';
+import productRoutes from './routes/productRoutes.js';
+import cors from 'cors';
+import connectDB from './config/db.js';
+
+const app = express();
+
+// configure env
+dotenv.config();
+
+//database config
+connectDB();
+
+//middlewares
+app.use(cors());
+app.use(express.json());
+app.use(morgan('dev'));
+
+//routes
+app.use('/api/v1/auth', authRoutes);
+app.use('/api/v1/category', categoryRoutes);
+app.use('/api/v1/product', productRoutes);
+
+// rest api
+
+app.get('/', (req, res) => {
+  res.send('<h1>Welcome to ecommerce app</h1>');
+});
+
+export default app;
