+++ server.js
@@ -1,41 +1,13 @@
-import express from "express";
-import colors from "colors";
-import dotenv from "dotenv";
-import morgan from "morgan";
-import connectDB from "./config/db.js";
-import authRoutes from './routes/authRoute.js'
-import categoryRoutes from './routes/categoryRoutes.js'
-import productRoutes from './routes/productRoutes.js'
-import orderRoutes from './routes/orderRoutes.js'
-import cors from "cors";
+import dotenv from 'dotenv';
+import app from './app.js';
 
 // configure env
 dotenv.config();
 
-//database config
-connectDB();
-
-const app = express();
-
-//middlewares
-app.use(cors());
-app.use(express.json());
-app.use(morgan('dev'));
-
-//routes
-app.use("/api/v1/auth", authRoutes);
-app.use("/api/v1/category", categoryRoutes);
-app.use("/api/v1/product", productRoutes);
-app.use("/api/v1/order", orderRoutes);
-
-// rest api
-
-app.get('/', (req,res) => {
-    res.send("<h1>Welcome to ecommerce app</h1>");
-});
-
 const PORT = process.env.PORT || 6060;
 
 app.listen(PORT, () => {
-    console.log(`Server running on ${process.env.DEV_MODE} mode on ${PORT}`.bgCyan.white);
+  console.log(
+    `Server running on ${process.env.DEV_MODE} mode on ${PORT}`.bgCyan.white
+  );
 });
